+++ supersql/__init__.py
@@ -1,4 +1,4 @@
 from supersql.core import Query, Table
 
 
-__version__ = "2026.2.1"
+__version__ = "2026.2.6"